requirements.txt
Django==5.1.7
djangorestframework-simplejwt==5.5.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
celery[redis]==5.4.0
django-celery-results==2.5.1
gunicorn==23.0.0
uvicorn[standard]==0.34.0
cloudinary==1.44.1
